requirements.txt
Flask
pytest
pandas
openpyxl
requests
google-generativeai
